fix(donate): show whole-person impact estimates
Round donor-facing impacted survivors to whole people and update donate-page copy so impact messaging stays realistic, clear, and encouraging for smaller gifts.

Made-with: Cursor

+++ backend/Intex.Api/DTOs/DonationDtos.cs
@@ -77,7 +77,7 @@ public record DonationImpactPredictionResponse(
     decimal Amount,
     IReadOnlyList<DonationImpactPredictionOutcomeResponse> Outcomes,
     string Assumptions,
-    decimal EstimatedVictimsImpacted);
+    int EstimatedVictimsImpacted);
 
 public record PublicDonationSubmissionRequest(
     bool IsAnonymous,
